Move queries / mutations into their own folder
