+++ packages/apollos-ui-auth/src/ProtectedAction.js
@@ -2,7 +2,7 @@ import React, { PureComponent } from 'react';
 import { Query } from 'react-apollo';
 import PropTypes from 'prop-types';
 
-import GET_LOGIN_STATE from './getLoginState';
+import { GET_LOGIN_STATE } from './queries';
 
 import { AuthConsumer } from './Provider';
