+++ packages/apollos-ui-auth/src/mutations/index.js
@@ -1,0 +1,7 @@
+export AUTHENTICATE from './authenticate';
+export HANDLE_LOGIN from './handleLogin';
+export LOGOUT from './logout';
+export REGISTER_WITH_SMS from './registerWithSms';
+export REGISTER_WITH_EMAIL from './registerWithEmail';
+export REQUEST_SMS_PIN from './requestPin';
+export VERIFY_PIN from './verifyPin';
