+++ packages/apollos-ui-auth/src/mutations/logout.js
