+++ packages/apollos-ui-auth/src/queries/getLoginState.js
