+++ packages/apollos-ui-auth/src/queries/index.js
@@ -1,0 +1,2 @@
+export GET_LOGIN_STATE from './getLoginState';
+export GET_USER_EXISTS from './getUserExists';
